lab04.py
The purpose of the program is to calculate the year when a person will turn 100 based on their current age or year of birth. It allows users to  input a specific future year and calculates the age they would be in that year. The program is designed to handle various types of errors and cases, ensuring a robust and user-friendly experience.​